Fix Gtk+ crash on GtkFontChooser initialization
When parsing the treemodel in gtkfontchooserwidget.ui, all the types used
there need to be defined. When Gtk+ encounters an undefined type, it tries
to initialize it itself by looking up the *_get_type() function and calling
it. This works fine when that symbol is defined in the main application
itself. However the QGtk3Theme is a plugin, so Gtk+ was unable to find the
type and was crashing later.

To fix the crash, initialize the two Pango types (PangoFontFamily and
PangoFontFace) in QGtk3Theme constructor. Also add some asserts to prevent
GCC from optimizing out the pango_*_get_type() calls.